Testing environments defaults
This sets some defaults for the test environment, namely configuring
RSpec options to disable monkey patching and setting up support gems.

As for support gems, this PR adds Webmock and the default config for it.

It also follows a convention of moving `require` statements and any
supporting gem configuration to a support file, named after the
configured gem. This makes it simple to find/add/modify configuration
for each support gem without traversing an overly long `rails_helper` or
`spec_helper` file.