Add change links to summary view
This adds "Change" links to the summary view, allowing building managers
to edit their answers before submitting the survey.

This makes the assumption that we are not allowing editing from anywhere
else in the application other than the summary view - hence all "update"
actions redirect back to the summary view.