Add Procfiles
This adds a staging/production Procfile as well as a development
Procfile (Procfile.dev).

For development, this allows starting the application with `foreman
start -f Procfile.dev´, which will boot the rails server and the webpack
server.

You can also boot the application in development with `heroku local -f
Procfile.dev` to enjoy the same benefits.

For the former, the `foreman` gem should be installed in the dev machine
and for the latter, the Heroku CLI tool should be installed.

This also adjust the puma config to run with Workers and preloading the
app.