Login to admin area
In order to enter data and manage sending out surveys, an admin area was
required. This should also be password protected and have no links
accessible publicly.

This PR adds an admin dashboard, secured by email/password, and
accessible via "/admin" or "/admin/dashboard".

In order to display navigation to the admin user, the application header
was modified to include links, only displayable if a user is presently
signed in.