Collapse building manager into building
The building manager record, or entity, has become redundant in the
project. It was originally created with the intention of allowing
communication of several surveys/buildings to a single entity.

As the project developed, it has now become apparent that we will not be
addressing a single entity that owns many buildings and presenting a
list of surveys to complete.

On the other hand, now that we have finalized data to work with, having
a separate building manager entity makes the data import a two stage
process: we'd need to import all buildings and then re-run the import
in order to import building managers.

As such, it is a great example of the example "Make the change easy, then
make an easy change": by collapsing the building manager data into the
building table, we make it easier to manage data via importing, or even
via a CRUD interface, as we are only handling one model.

A small side effect is an accumulation of data in the building model,
but even then it doesn't strike me at this point as worth the trade-off.
It seems better to be able to ask a building "What is email address for
your proprietor?" than making any data import much more complex.