Allow admin to send survey link via letter
This adds a structure for sending notifications to potential survey
users. It implements sending letter notifications and reuses the work
done to integrate with the Notify API.

This also implements a somewhat simplistic system for parsing addresses,
based on the presence of an UK postcode. The objective is to ensure the
last line of address passed to Notify contains either a valid UK
postcode or a foreign country.