form: use a callback to set the survey as completed
This approach allows all stages to run some logic once the form for a
particular stage is instanciated and the record (with the relevant
data) attached to it.

In the case of the "check your answers" stage, this allows us to mark
the survey as completed when users reach it, and then send them back
to it if they decide to amend a field.