csv: correctly cast other values for material/insulations
We were assuming that `material_X` and `insulation_X` should stay as
known values from the list of options that can be selected, including
"other".

However DELTA and its CSV cast the associated text input to that
column when "Other" is selected so that if you enter something like
"Feathers", it moves into that column along with the other predefined
options.

The confusion might have appeared because DELTA does it differently
for balconies and external structures by explicitly providing columns
like `shading-materials-other` for that purpose.